requirements.txt
# --- 基础工具 ---
tqdm>=4.66.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0

# --- 智能体与API (Teacher Agent) ---
openai>=1.10.0

# --- 模型推理 (Student Model) ---
torch>=2.1.0
transformers>=4.37.0
peft>=0.7.0
accelerate>=0.26.0
bitsandbytes>=0.41.0

# --- 前端界面 ---
gradio>=4.0.0